#3 fix: Bug Fixes
