+++ src/configs/QueryRules.ts
