--- src/configs/schemas/queryRequest.json
@@ -1,117 +1,0 @@
-{
-  "type": "object",
-  "properties": {
-    "query": {
-      "type": "object",
-      "properties": {
-        "queryType": { "type": "string" },
-        "dataSource": { "type": "string" },
-        "intervals": {
-          "oneOf": [
-            {
-              "type": "string"
-            },
-            {
-              "type": "array",
-              "items": {
-                "type": "string"
-              }
-            }
-          ]
-        },
-        "granularity": {
-          "oneOf": [
-            {
-              "type": "string"
-            },
-            {
-              "type": "object",
-              "required": []
-            }
-          ]
-        },
-        "dimension": {
-          "oneOf": [
-            {
-              "type": "string"
-            },
-            {
-              "type": "object",
-              "required": []
-            }
-          ]
-        },
-        "dimensions": {
-          "type": "array",
-          "items": {
-            "oneOf": [
-              {
-                "type": "string"
-              },
-              {
-                "type": "object",
-                "required": []
-              }
-            ]
-          },
-          "nullable": true
-        },
-        "filter": {
-          "type": "object",
-          "required": [],
-          "nullable": true
-        },
-        "aggregations": {
-          "type": "array",
-          "items": {
-            "type": "object",
-            "required": [],
-            "nullable": true
-          },
-          "nullable": true
-        },
-        "postAggregations": {
-          "type": "array",
-          "items": {
-            "type": "object",
-            "required": [],
-            "nullable": true
-          },
-          "nullable": true
-        },
-        "threshold": {
-          "type": "number",
-          "nullable": true
-        },
-        "limit": { "type": "number", "nullable": true }
-      },
-      "required": [],
-      "nullable": true
-    },
-    "querySql": {
-      "type": "object",
-      "properties": {
-        "query": {
-          "type": "string"
-        }
-      },
-      "required": ["query"],
-      "nullable": true
-    },
-    "context": {
-      "type": "object",
-      "properties": {
-        "dataSource": {
-          "type": "string"
-        }
-      },
-      "required": ["dataSource"]
-    }
-  },
-  "oneOf": [
-    {
-      "required": ["context", "querySql"]
-    },
-    { "required": ["context", "query"] }
-  ]
-}
